+++ src/ComputeSharp.Graphics/Buffers/Abstract/StructuredBuffer{T}.cs
@@ -28,7 +28,7 @@ namespace ComputeSharp.Graphics.Buffers.Abstract
         }
 
         /// <inheritdoc/>
-        internal override unsafe void GetData(ref T destination, nint size, int offset)
+        internal override unsafe void GetData(ref T destination, int size, int offset)
         {
             GraphicsDevice.ThrowIfDisposed();
 
@@ -80,7 +80,7 @@ namespace ComputeSharp.Graphics.Buffers.Abstract
         }
 
         /// <inheritdoc/>
-        internal override unsafe void SetData(ref T source, nint size, int offset)
+        internal override unsafe void SetData(ref T source, int size, int offset)
         {
             GraphicsDevice.ThrowIfDisposed();
