+++ src/ComputeSharp.Graphics/Buffers/Abstract/Texture3D{T}.cs
@@ -138,7 +138,7 @@ namespace ComputeSharp.Graphics.Buffers.Abstract
         /// <param name="width">The width of the memory area to copy.</param>
         /// <param name="height">The height of the memory area to copy.</param>
         /// <param name="depth">The depth of the memory area to copy.</param>
-        internal void GetData(ref T destination, nint size, int x, int y, int z, int width, int height, int depth)
+        internal void GetData(ref T destination, int size, int x, int y, int z, int width, int height, int depth)
         {
             GraphicsDevice.ThrowIfDisposed();
 
@@ -217,7 +217,7 @@ namespace ComputeSharp.Graphics.Buffers.Abstract
         /// <param name="width">The width of the memory area to write to.</param>
         /// <param name="height">The height of the memory area to write to.</param>
         /// <param name="depth">The depth of the memory area to write to.</param>
-        internal void SetData(ref T source, nint size, int x, int y, int z, int width, int height, int depth)
+        internal void SetData(ref T source, int size, int x, int y, int z, int width, int height, int depth)
         {
             GraphicsDevice.ThrowIfDisposed();
